Support mixed MAPADDRESS response
Recently tor documented that MAPADDRESS can contain a mixture of successful and
failed responses...

  https://gitweb.torproject.org/torspec.git/commit/?id=1417af05de247d9f858863849d16a7185577d369

Such a response didn't break us, our map_address() method simply ignored the
failures and returned a dictionary with whatever mappings succeeded.

However, we should provide our caller with failures as well for completeness.
This breaks backward compatibility in a couple ways...

  1. Our map_address() method now returns a MapAddressResponse rather
     than a dictionary.

  2. Renamed MapAddressResponse's "entries" attribute to "mappings".

Personally I'd prefer if MAPADDRESS was all-or-nothing (that is to say, if any
entry is malformed then reject the request). That would produce a simpler API
for our callers. But I don't control that so this is the method response we'll
need to supply.