Test that the Query class' 'start' argument works
Reading our Query class I became worried that our 'start' argument fails to get
honored because the start() method is asynchronous (so its invocation returns a
coroutine rather than running it).

I was wrong. It works because as a Synchronous subclass our metaprogramming
converts the call. That said, this is none the less a good thing to check.